refactor(server): extract /channels route group (Refs #3502)
Move the 11 inline `/channels*` handlers out of server.ts into a new
dedicated `src/server/routes/channelRoutes.ts`, mounted via
`apiRouter.use('/channels', channelRoutes)`. Handler bodies are moved
verbatim (paths made relative to the `/channels` prefix); auth/validation
middleware (optionalAuth / requireAuth / requirePermission) is preserved
identically and in order, so the final API paths are byte-identical.

Handlers moved (11): GET /channels/all, GET /channels, GET
/channels/:id/export, PUT /channels/:id, DELETE /channels/:id, POST
/channels/:slotId/import, POST /channels/reorder, POST
/channels/decode-url, POST /channels/encode-url, POST
/channels/import-config, POST /channels/refresh.

The three channel-move helpers (detectChannelMoves,
snapshotChannelsBeforeChange, migrateMessagesIfChannelsMoved) were
server.ts-local and used only by these handlers, so they moved into the
new module too. Their now-orphaned server.ts imports (migrateAutomationChannels,
modemPresetChannelName) are removed.

Inline `apiRouter.{get,post,put,delete,patch}` handler count in server.ts:
89 -> 78.

Adds routes/channelRoutes.test.ts (31 cases) covering CRUD validation +
happy paths, import/export validation, URL encode/decode, reorder
permutation checks, and MeshCore branch guards.
